refactor: setting, close icon svg로 교체
