+++ src/components/common/TabIcons.tsx
@@ -5,8 +5,8 @@ import CalendarIconSVG from '@assets/images/diary.svg';
 import PieceIconSVG from '@assets/images/piece.svg';
 import MailOpen from '@assets/svg/icons/ion--mail-open-outline.svg';
 import Mail from '@assets/svg/icons/ion--mail-outline.svg';
-import AntDesign from 'react-native-vector-icons/AntDesign';
-import MyIconButtons from '@components/common/MyIconButtons';
+import Shop from '@assets/svg/icons/tdesign--shop-5.svg';
+import MyIconButton from '@components/common/MyIconButton';
 import { useRecoilValue } from 'recoil';
 import { userInfoState } from '@stores/login';
 import { alertColor } from '@utils/colors';
@@ -34,7 +34,7 @@ export const AiLetterIcon = ({ focused, color, size }: ITabBarIconProps) => {
 };
 
 export const ShopIcon = memo(({ color }: ITabBarIconProps) => (
-  <AntDesign name="isv" color={color} size={24.5} />
+  <Shop width={24.5} height={24.5} fill={color} />
 ));
 
 export const PieceIcon = memo(({ color }: ITabBarIconProps) => {
@@ -42,13 +42,7 @@ export const PieceIcon = memo(({ color }: ITabBarIconProps) => {
 });
 
 export const CloseIcon = ({ onPress }: { onPress: () => {} }) => (
-  <MyIconButtons
-    name={'close'}
-    iconSet={'MaterialCommunityIcons'}
-    size={24}
-    onPress={onPress}
-    style={{ paddingHorizontal: 16 }}
-  />
+  <MyIconButton name="close" style={{ paddingHorizontal: 16 }} onPress={onPress} />
 );
 
 const styles = StyleSheet.create({
